Simplify analytics routing
