+++ client/web/src/site-admin/routes.tsx
@@ -4,6 +4,46 @@ import { SiteAdminAreaRoute } from './SiteAdminArea'
 
 export const siteAdminAreaRoutes: readonly SiteAdminAreaRoute[] = [
     {
+        path: '/',
+        render: lazyComponent(() => import('./analytics/AnalyticsOverviewPage'), 'AnalyticsOverviewPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/search',
+        render: lazyComponent(() => import('./analytics/AnalyticsSearchPage'), 'AnalyticsSearchPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/code-intel',
+        render: lazyComponent(() => import('./analytics/AnalyticsCodeIntelPage'), 'AnalyticsCodeIntelPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/extensions',
+        render: lazyComponent(() => import('./analytics/AnalyticsExtensionsPage'), 'AnalyticsExtensionsPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/users',
+        render: lazyComponent(() => import('./analytics/AnalyticsUsersPage'), 'AnalyticsUsersPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/code-insights',
+        render: lazyComponent(() => import('./analytics/AnalyticsCodeInsightsPage'), 'AnalyticsCodeInsightsPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/batch-changes',
+        render: lazyComponent(() => import('./analytics/AnalyticsBatchChangesPage'), 'AnalyticsBatchChangesPage'),
+        exact: true,
+    },
+    {
+        path: '/analytics/notebooks',
+        render: lazyComponent(() => import('./analytics/AnalyticsNotebooksPage'), 'AnalyticsNotebooksPage'),
+        exact: true,
+    },
+    {
         path: '/configuration',
         exact: true,
         render: lazyComponent(() => import('./SiteAdminConfigurationPage'), 'SiteAdminConfigurationPage'),
